Add package dtos for tranfer data access and fix controller, service

+++ src/main/java/goldiounes/com/vn/services/ProductService.java
@@ -1,41 +1,95 @@
 package goldiounes.com.vn.services;
 
-import goldiounes.com.vn.models.Product;
+import goldiounes.com.vn.models.dto.ProductDTO;
+import goldiounes.com.vn.models.entity.Category;
+import goldiounes.com.vn.models.entity.Product;
+import goldiounes.com.vn.repositories.CategoryRepo;
 import goldiounes.com.vn.repositories.ProductRepo;
+import org.modelmapper.ModelMapper;
+import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
+
     @Autowired
     private ProductRepo productRepo;
 
-    public List<Product> findAll() {
-        return productRepo.findAll();
+    @Autowired
+    private CategoryRepo categoryRepo;
+
+    @Autowired
+    private ModelMapper modelMapper;
+
+    public List<ProductDTO> getAllProducts() {
+        List<Product> products = productRepo.findAll();
+        if (products.isEmpty()) {
+            throw new RuntimeException("No products found");
+        }
+        return modelMapper.map(products, new TypeToken<List<ProductDTO>>() {}.getType());
     }
 
-    public Product findById(int id) {
-        return productRepo.findById(id).get();
+    public ProductDTO getProduct(int id) {
+        Optional<Product> existingProduct = productRepo.findById(id);
+        if (existingProduct.isEmpty()) {
+            throw new RuntimeException("Product not found");
+        }
+        return modelMapper.map(existingProduct.get(), ProductDTO.class);
     }
 
-    public Product save(Product product) {
-        return productRepo.save(product);
-    }
-    public void delete(Product product) {
-        productRepo.delete(product);
+    public ProductDTO createProduct(ProductDTO productDTO) {
+        Optional<Product> existingProduct = productRepo.findById(productDTO.getProductID());
+        if (existingProduct.isPresent()) {
+            throw new RuntimeException("Product already exists");
+        }
+        Category category = categoryRepo.findByName(productDTO.getCategory().getCategoryName());
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
+        Product product = modelMapper.map(productDTO, Product.class);
+        product.setCategory(category);
+        productRepo.save(product);
+        return modelMapper.map(product, ProductDTO.class);
     }
 
-    public List<Product> findByName(String name) {
-        return productRepo.findByProductName(name);
+    public void deleteProduct(int id) {
+        Optional<Product> existingProduct = productRepo.findById(id);
+        if (existingProduct.isEmpty()) {
+            throw new RuntimeException("Product not found");
+        }
+        productRepo.delete(existingProduct.get());
+    }
+
+    public List<ProductDTO> findByName(String name) {
+        List<Product> products = productRepo.findByProductName(name);
+        return modelMapper.map(products, new TypeToken<List<ProductDTO>>() {}.getType());
     }
 
     public void deleteById(int id) {
+        if (!productRepo.existsById(id)) {
+            throw new RuntimeException("Product not found");
+        }
         productRepo.deleteById(id);
     }
 
-    public List<Product> findByCategory(String KeyWord) {
-        return productRepo.findByCategoryId(KeyWord);
+    public List<ProductDTO> findByCategory(String keyWord) {
+        List<Product> products = productRepo.findByCategoryKeyWord(keyWord);
+        if (products.isEmpty()) {
+            throw new RuntimeException("No products found");
+        }
+        return modelMapper.map(products, new TypeToken<List<ProductDTO>>() {}.getType());
     }
+
+    public ProductDTO updateProduct(int id, ProductDTO productDTO) {
+        Product existingProduct = productRepo.findById(id).get();
+        if (existingProduct == null) {
+            throw new RuntimeException("Product not found");
+        }
+        return modelMapper.map(productRepo.save(existingProduct), ProductDTO.class);
+    }
+
 }
